+++ RELEASE_NOTES.md
@@ -5,6 +5,7 @@
  * #9: New command `Map Preview (with projection)`, allowing for data to be previewed with a declared projection (required for formats like WKT, that have no inferred geographic projection)
  * #10: Fix subsequent previews not working properly
  * #11: Configuration option to allow preview content to be dumped out to an external HTML file (for debugging purposes)
+ * #15: Default base layer is now configurable
  * Some small CSS updates/fixes
 
 # 0.2.2